moving general tests in lstchain (where they should have been from start) and fixing bug for some other tests
