+++ lstchain/mc/tests/test_senstivity.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pytest
 import pandas as pd
-from ctapipe.utils import get_dataset_path 
 from lstchain.mc import (
     read_sim_par,
     calculate_sensitivity,
@@ -10,10 +9,11 @@ from lstchain.mc import (
     ring_containment,
 )
 
+from lstchain.tests.test_lstchain import dl1_file, dl2_file, lstcam_key
+
 
 @pytest.mark.run(after='test_dl0_to_dl1')
 def test_read_sim_par():
-    dl1_file = 'dl1_gamma_test_large.h5'
     par = read_sim_par(dl1_file)
 
     assert np.isclose(par['emin'].to_value(), 0.003)
@@ -26,9 +26,9 @@ def test_read_sim_par():
 
 @pytest.mark.run(after='test_apply_models')
 def test_process_mc():
-    # need to define a dl2 test_dataset
-    pd.read_hdf('dl2_gamma_test_large.h5', key='events/LSTCam')
-
+    # TODO: define a dl2 test_dataset
+    # pd.read_hdf(dl2_file, key=lstcam_key)
+    pass
 
 def test_calculate_sensitivity():
 
@@ -94,7 +94,9 @@ def test_ring_containment():
 
 @pytest.mark.run(after='test_apply_models')
 def test_sens():
-    dl2 = pd.read_hdf('dl2_gamma_test_large.h5', key='events/LSTCam')
+    # TODO: define test for sens
+    # dl2 = pd.read_hdf(dl2_file, key=lstcam_key)
+    pass
